requirements.txt
PySide6>=6.5
openai-whisper
torch
pyaudio
pynput
pyperclip
requests
Pillow
